+++ source/common/ipc/api.js
@@ -183,3 +183,9 @@ export const SET_CACHED_CARDANO_STATUS_CHANNEL =
   'SET_CACHED_CARDANO_STATUS_CHANNEL';
 export type SetCachedCardanoStatusRequest = ?CardanoStatus;
 export type SetCachedCardanoStatusResponse = void;
+
+/**
+ * Channel where renderer can ask main process for the result of electron's app.getLocale()
+ */
+export const DETECT_SYSTEM_LOCALE_CHANNEL = 'DETECT_SYSTEM_LOCALE_CHANNEL';
+export type DetectSystemLocaleResponse = string;