+++ source/common/types/locales.types.js
@@ -1,0 +1,10 @@
+// @flow
+export type Locales = {
+  english: string,
+  japanese: string,
+};
+
+export const LOCALES: Locales = {
+  english: 'en-US',
+  japanese: 'ja-JP',
+};