+++ source/renderer/app/components/wallet/transactions/WalletTransactionsList.js
@@ -281,6 +281,7 @@ export default class WalletTransactionsList extends Component<Props> {
       <Button
         className={buttonClasses}
         label={intl.formatMessage(messages.showMoreTransactionsButtonLabel)}
+        // eslint-disable-next-line react/jsx-no-bind
         onClick={this.onShowMoreTransactions.bind(this, walletId)}
         skin={ButtonSkin}
       />
@@ -292,7 +293,9 @@ export default class WalletTransactionsList extends Component<Props> {
         {isRenderingAsVirtualList ? (
           <VirtualTransactionList
             getExpandedTransactions={this.getExpandedTransactions}
-            ref={list => (this.virtualList = list)}
+            ref={list => {
+              this.virtualList = list;
+            }}
             renderRow={this.renderItem}
             rows={rows}
             isLoadingSpinnerShown={loadingSpinner !== null}
@@ -300,7 +303,9 @@ export default class WalletTransactionsList extends Component<Props> {
           />
         ) : (
           <SimpleTransactionList
-            ref={list => (this.simpleList = list)}
+            ref={list => {
+              this.simpleList = list;
+            }}
             renderRow={this.renderItem}
             rows={rows}
           />