+++ source/renderer/app/utils/crypto.js
@@ -71,7 +71,7 @@ export const unscramblePaperWalletMnemonic = (
 
 export const mnemonicToSeedHex = (mnemonic: string, password: ?string) => {
   const mnemonicBuffer = Buffer.from(unorm.nfkd(mnemonic), 'utf8');
-  const salt = 'mnemonic' + (unorm.nfkd(password) || '');
+  const salt = `mnemonic${unorm.nfkd(password) || ''}`;
   const saltBuffer = Buffer.from(salt, 'utf8');
   return pbkdf2(mnemonicBuffer, saltBuffer, 2048, 32, 'sha512').toString('hex');
 };